+++ src/features/TextToImage/webAPI/Server/index.ts
@@ -8,7 +8,7 @@ import {
   DynamicConfig,
   StaticConfig,
   emptyConfig,
-} from '../../config';
+} from '../../config/exports';
 
 import TextToImage_Server_SpecificationError from './ServerSpecificationError';
